# docs/swagger.yaml
definitions:
...
  models.CreateProgramadReq:
    properties:
      desde:
        type: integer
      hasta:
        type: integer
      id:
        type: string
      liberado:
        type: integer
      programa_id:
        type: integer
      valor:
        type: number
    type: object
  models.CreateProgramsReq:
    properties:
      active:
        type: integer
      author:
        type: string
      code:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      destination:
        type: string
      destination_code:
        type: string
      id:
        type: string
      matrix:
        type: object
      name:
        type: string
      origin:
        type: string
      origin_code:
        type: string
      reserva:
        type: number
      updateDate:
        type: string
    type: object
...
  models.ProgramadResp:
    properties:
      desde:
        type: integer
      hasta:
        type: integer
      id:
        type: string
      liberado:
        type: integer
      programa_id:
        type: integer
      valor:
        type: number
    type: object
...
  models.ProgramsResp:
    properties:
      active:
        type: integer
      author:
        type: string
      code:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      destination:
        type: string
      destination_code:
        type: string
      id:
        type: string
      matrix:
        type: object
      name:
        type: string
      origin:
        type: string
      origin_code:
        type: string
      reserva:
        type: number
      updateDate:
        type: string
    type: object
...
  models.UpdateProgramsReq:
    properties:
      active:
        type: integer
      author:
        type: string
      code:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      destination:
        type: string
      destination_code:
        type: string
      matrix:
        type: object
      name:
        type: string
      origin:
        type: string
      origin_code:
        type: string
      reserva:
        type: number
      updateDate:
        type: string
    type: object
...
info:
  contact: {}
  description: Powered by scv-go-tools - https://github.com/sergicanet9/scv-go-tools
  title: Go Hexagonal API
paths:
...
  /api/v3.5/program:
    get:
      description: Gets all the program
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProgramsResp'
            type: array
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get all program
      tags:
      - program
    post:
      description: Creates a new program
      parameters:
      - description: New program to be created
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateProgramsReq'
      responses:
        "201":
          description: OK
          schema:
            $ref: '#/definitions/models.ProgramsResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Create program
      tags:
      - program
  /api/v3.5/program/{id}:
    delete:
      description: Delete a program
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Delete program
      tags:
      - program
    get:
      description: Gets a program by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProgramsResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get program by ID
      tags:
      - program
    patch:
      description: Updates a program
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: program
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProgramsReq'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Update program
      tags:
      - program
...
  /v2.0/program:
    post:
      description: Creates a new program
      parameters:
      - description: New program to be created
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateProgramadReq'
      responses:
        "201":
          description: OK
          schema:
            $ref: '#/definitions/models.ProgramadResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Create program
      tags:
      - program
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"